Add rollback functionality.
Explanation: We are making a new job in concourse called rollback-jobs that will pass in a param "rollback: true" to this script. If this condition is met, this script will call the rollback method, which calls "kubectl undo deployment" on that resource. Note that non-rollback jobs from concourse do not need to pass in "rollback: false" since we are only checking for the true condition to be met.